Move the vote entity under the lib namespace
