+++ lib/cast_vote_msg.js
@@ -1,5 +1,5 @@
 const Assert = require('assert-plus')
-const Vote = require('./entities/sys/vote')
+const Vote = require('./vote_entity')
 const Shapes = require('./shapes')
 const { fetchProp } = require('./utils')
 const { ValidationError, NotFoundError } = require('./errors')
