+++ lib/get_vote_stats.js
@@ -1,5 +1,5 @@
 const Assert = require('assert-plus')
-const Vote = require('./entities/sys/vote')
+const Vote = require('./vote_entity')
 const groupBy = require('lodash.groupby')
 const { fetchProp, countMatching } = require('./utils')
