+++ lib/open_poll_msg.js
@@ -3,7 +3,7 @@ const Joi = require('joi')
 const Shapes = require('./shapes')
 const { fetchProp } = require('./utils')
 const { ValidationError } = require('./errors')
-const Vote = require('./entities/sys/vote')
+const Vote = require('./vote_entity')
 const Reply = require('./reply')
 
 module.exports = function (opts = {}) {
