+++ lib/vote_entity.js
@@ -1,5 +1,5 @@
 const Assert = require('assert-plus')
-const { fetchProp } = require('../../utils')
+const { fetchProp } = require('./utils')
 
 class Vote {
   static TYPE_UP() {
